Move method to helpers and rename methods
